requirements.txt
# Database drivers
cassandra-driver==3.25.0  # For ScyllaDB
duckdb==0.8.1

# Data processing
pandas==2.0.3
numpy==1.24.3
pyarrow==12.0.1

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Utilities
tqdm==4.65.0
pyyaml==6.0
python-dotenv==1.0.0

# Testing and benchmarking
pytest==7.3.1
pytest-benchmark==4.0.0

# Documentation
jupyter==1.0.0
nbconvert==7.3.1 